TypeScript support
