+++ src/components/BlogFullArticle.astro
@@ -1,6 +1,6 @@
 ---
 import { Image } from "astro:assets";
-import { formatDate } from "@utils/utils.js";
+import { formatDate } from "@utils/utils.ts";
 import "@styles/markdown.less";
 
 // Receiving incoming props and destructuring them
