+++ src/components/BlogPreview.astro
@@ -1,6 +1,6 @@
 ---
 import { Image } from "astro:assets";
-import { formatDate } from "@utils/utils.js";
+import { formatDate } from "@utils/utils.ts";
 import { getLocalePath, t } from "i18n:astro";
 import { handleI18nSlug } from "@astrolicious/i18n/content-collections";
