+++ src/components/FeaturedPost.astro
@@ -1,7 +1,7 @@
 ---
 import { getCollection } from "astro:content";
 import { Image } from "astro:assets";
-import { slugify, formatDate } from "@utils/utils.js";
+import { slugify, formatDate } from "@utils/utils.ts";
 import { getLocale, getLocalePath, t } from "i18n:astro";
 
 // Gets posts by locale AND by tag
